requirements.txt
requests==2.31.0
urllib3==2.0.4
mpmath==1.3.0
networkx==3.1
opencv-contrib-python==4.8.0.76
pandas>=1.1.4
seaborn>=0.11.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
smmap==5.0.0
tzdata==2023.3
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.22.2
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.8.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.9.0
tqdm>=4.64.0
ultralytics>=8.0.147